Move container/stream/bytespipe to daemon/internal/stream/bytespipe
