+++ daemon/internal/stream/bytespipe/buffer.go
