+++ daemon/internal/stream/bytespipe/bytespipe.go
