+++ daemon/internal/stream/bytespipe/bytespipe_test.go
